moved out MovieDetailed to a separate component
