+++ code/src/components/MovieListed.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-export const Movie = ({ id, title, poster_path }) => {
+export const MovieListed = ({ id, title, poster_path }) => {
   return (
     <Link className="movie-card-link" to={`/movies/${id}`}>
       {/* "to" decides the path name and the value of the movieId variable I have in App.js and MovieDetails */}
